Modularize code into screens
